separateLogic
