+++ src/components/UserProfile/UserReviews/ReviewedProducts/ReviewedProducts.jsx
@@ -1,0 +1,71 @@
+import styles from "./ReviewedProducts.module.css";
+import React, { useState, useEffect } from "react";
+import HopPassionClient from "../../../../utils/NetworkingUtils";
+import { useParams } from "react-router-dom";
+import { Star, StarFill } from "react-bootstrap-icons";
+import { Spinner } from "react-bootstrap";
+
+function ReviewedProducts() {
+  const [reviewedProducts, setReviewedProducts] = useState([]);
+  const [isLoading, setIsLoading] = useState(true); // Agregamos el estado de carga
+  const { id } = useParams();
+
+  const fetchReviewedProducts = async () => {
+    try {
+      const response = await HopPassionClient.get(`/product/qualified/${id}`);
+      setReviewedProducts(response.data);
+      setIsLoading(false); // Cambiamos isLoading a falso cuando se completó la carga
+    } catch (error) {
+      console.error("Error al obtener productos reseñados", error);
+      setIsLoading(false); // Aseguramos que isLoading sea falso en caso de error
+    }
+  };
+
+  function calculateAverageRating(reviews) {
+    if (reviews.length === 0) return 0;
+
+    const totalRating = reviews.reduce(
+      (total, review) => total + review.rating,
+      0
+    );
+    return totalRating / reviews.length;
+  }
+
+  useEffect(() => {
+    fetchReviewedProducts();
+  }, []);
+
+  return (
+    <div>
+      {isLoading ? (
+        <Spinner animation="border" role="status" />
+      ) : (
+        reviewedProducts.map((product) => (
+          <div key={product.id} className={styles.product}>
+            <div className={styles.productHeader}>
+              <h2 className={styles.productName}>{product.name}</h2>
+              <div className={styles.rating}>
+                {[1, 2, 3, 4, 5].map((star) => (
+                  <span key={star}>
+                    {star <= calculateAverageRating(product.Reviews) ? (
+                      <StarFill className={styles.starFill} />
+                    ) : (
+                      <Star className={styles.starOutline} />
+                    )}
+                  </span>
+                ))}
+              </div>
+            </div>
+            {product.Reviews.map((review, index) => (
+              <div key={index} className={styles.reviewItem}>
+                <p className={styles.productReview}>{review.comment}</p>
+              </div>
+            ))}
+          </div>
+        ))
+      )}
+    </div>
+  );
+}
+
+export default ReviewedProducts;
